requirements.txt
Flask
openai
PyMuPDF
python-dotenv
